Bring three tests up to the contracts the code now actually has
sip1OutputReady emits MIC-rate frames since the monitor decimation —
the frame-count test still expected the DUC-rate count, i.e. the exact
bug the bench heard as crackling. It now asserts 256 from a 1024-frame
P2 output block and documents why.

tst_waterfall_color still lived before the palette fixes of Aug 7: it
compared peaks against the REMOVED magenta cap (the real top is red
now — kClarityPeak), and its ramp helper sampled through a 1 dB window
that the collapsed-window guard silently widens, so 256 samples covered
the bottom sliver of the palette and CIELAB's cube-root steepness off
pure black read as a 'hard edge'. The helper now maps t straight to
palette position through a 100 dB window; the old magenta constant
stays as a tripwire against the cap returning.

tst_eq_filter_ring promoted stored floats to double before comparing —
0.9f is 0.8999999761… as a double and QCOMPARE's double epsilon
rightly refused it. Compared as float now, which is what the field is.

Plus test-failures.log gitignored — it is the tee target of the
debugging workflow, not a work product.

+++ tests/tst_eq_filter_ring.cpp
@@ -159,7 +159,11 @@ private slots:
         c->eq().setBand(4, p);
 
         Ring::applyTo(c->eq(), 4, FT::Peak);
-        QCOMPARE(double(c->eq().band(4).q), 0.9);
+        // Compare as FLOAT: q is stored as float, and 0.9 has no exact
+        // float representation — promoting to double manufactures a
+        // difference (0.8999999761…) that QCOMPARE's double-epsilon
+        // rejects. Bitten on the first full-suite run, 2026-08-11.
+        QCOMPARE(c->eq().band(4).q, 0.9f);
     }
 
     void the_clamp_only_ever_lowers_q()
@@ -172,7 +176,7 @@ private slots:
         c->eq().setBand(5, p);
 
         Ring::applyTo(c->eq(), 5, FT::HighPass);
-        QCOMPARE(double(c->eq().band(5).q), 0.3);
+        QCOMPARE(c->eq().band(5).q, 0.3f);  // float compare — see above
     }
 
     // Slopes ignore gain. If the write cleared it, a stray double-click
@@ -199,7 +203,7 @@ private slots:
         QCOMPARE(double(after.freqHz), 1750.0);
         // Q is the one thing a lap is allowed to change, because the lap
         // passes through four types that cannot hold a sharp one.
-        QCOMPARE(double(after.q), 0.8);
+        QCOMPARE(after.q, 0.8f);  // float compare — see above
     }
 };
